MINI- Project/projectdata/requirements.txt
python-barcode==0.11.0
